requirements.txt
anthropic>=0.25.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.25.0
